+++ DurableMultiAgentTemplate.Client/Components/Chat/AdditionalInfoList.razor
@@ -1,4 +1,4 @@
-﻿@using DurableMultiAgentTemplate.Model
+﻿@using DurableMultiAgentTemplate.Shared.Model
 @using Markdig
 <div style="width: 100%; height: 100%;">
     <h3>追加情報</h3>
