+++ DurableMultiAgentTemplate.Client/Components/Pages/Home.razor
@@ -2,7 +2,7 @@
 @using DurableMultiAgentTemplate.Client.Components.Chat
 @using DurableMultiAgentTemplate.Client.Components.Utilities
 @using DurableMultiAgentTemplate.Client.Model
-@using DurableMultiAgentTemplate.Model
+@using DurableMultiAgentTemplate.Shared.Model
 @using Markdig
 
 <PageTitle>Simple agent client</PageTitle>
