+++ DurableMultiAgentTemplate.Client/Components/Pages/Home.razor.cs
@@ -2,7 +2,8 @@
 using DurableMultiAgentTemplate.Client.Model;
 using DurableMultiAgentTemplate.Client.Services;
 using DurableMultiAgentTemplate.Client.Utilities;
-using DurableMultiAgentTemplate.Model;
+using DurableMultiAgentTemplate.Shared.Model;
+
 
 namespace DurableMultiAgentTemplate.Client.Components.Pages;
