+++ DurableMultiAgentTemplate.Client/Services/AgentChatService.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text.Json;
-using DurableMultiAgentTemplate.Model;
+using DurableMultiAgentTemplate.Shared.Model;
 
 namespace DurableMultiAgentTemplate.Client.Services;
