+++ DurableMultiAgentTemplate.Shared/Model/AdditionalLinkInfo.cs
@@ -1,0 +1,12 @@
+using System.ComponentModel;
+
+namespace DurableMultiAgentTemplate.Shared.Model;
+
+class AdditionalLinkInfo : IAdditionalInfo
+{
+    [Description("リンクのラベルとして表示されるテキスト")]
+    public string LinkText { get; set; } = "";
+
+    [Description("リンク先のURL")]
+    public required Uri Uri { get; set; }
+}
