+++ DurableMultiAgentTemplate.Shared/Model/AdditionalMarkdownInfo.cs
@@ -1,0 +1,9 @@
+using System.ComponentModel;
+
+namespace DurableMultiAgentTemplate.Shared.Model;
+
+class AdditionalMarkdownInfo : IAdditionalInfo
+{
+    [Description("Markdown形式の補足情報")]
+    public string MarkdownText { get; set; } = "";
+}
