+++ DurableMultiAgentTemplate.Shared/Model/AgentRequestDto.cs
@@ -1,0 +1,10 @@
+using System.ComponentModel;
+
+namespace DurableMultiAgentTemplate.Shared.Model;
+
+public class AgentRequestDto
+{
+    public List<AgentRequestMessageItem> Messages { get; set; } = default!;
+    
+    public bool RequireAdditionalInfo { get; set; } = false;
+}
