+++ DurableMultiAgentTemplate.Shared/Model/AgentRequestMessageItem.cs
@@ -1,0 +1,9 @@
+using System.ComponentModel;
+
+namespace DurableMultiAgentTemplate.Shared.Model;
+
+public class AgentRequestMessageItem
+{
+    public string Role { get; set; } = string.Empty;
+    public string Content { get; set; } = string.Empty;
+}
