+++ DurableMultiAgentTemplate/Agent/Orchestrator/AgentDeciderActivity.cs
@@ -6,6 +6,7 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using OpenAI.Chat;
+using DurableMultiAgentTemplate.Shared.Model;
 
 
 namespace DurableMultiAgentTemplate.Agent.Orchestrator;
