+++ DurableMultiAgentTemplate/Agent/Orchestrator/SynthesizerWithAdditionalInfoActivity.cs
@@ -1,13 +1,12 @@
 ﻿using System.Text.Json;
-using System.Text.Json.Serialization;
-using Azure.AI.OpenAI;
 using DurableMultiAgentTemplate.Extension;
 using DurableMultiAgentTemplate.Json;
 using DurableMultiAgentTemplate.Model;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using OpenAI.Chat;
+using DurableMultiAgentTemplate.Shared.Model;
+
 
 namespace DurableMultiAgentTemplate.Agent.Orchestrator;
