+++ DurableMultiAgentTemplate/DurableMultiAgentTemplate.csproj
@@ -35,6 +35,11 @@
   <ItemGroup>
     <Using Include="System.Threading.ExecutionContext" Alias="ExecutionContext" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\DurableMultiAgentTemplate.Shared\Model\*.cs" LinkBase="Model" />
+  </ItemGroup>
+
   <Target Name="CopyGrpcNativeAssetsToOutDir" AfterTargets="Build">
     <ItemGroup>
       <NativeAssetToCopy Condition="$([MSBuild]::IsOSPlatform('OSX'))" Include="$(OutDir)runtimes/osx-arm64/native/*" />
