--- DurableMultiAgentTemplate/Model/AgentRequestDto.cs
@@ -1,14 +1,0 @@
-namespace DurableMultiAgentTemplate.Model;
-
-public class AgentRequestDto
-{
-    public List<AgentRequestMessageItem> Messages { get; set; } = default!;
-    
-    public bool RequireAdditionalInfo { get; set; } = false;
-}
-
-public class AgentRequestMessageItem
-{
-    public string Role { get; set; } = string.Empty;
-    public string Content { get; set; } = string.Empty;
-}
